+++ src/main/java/com/example/gymerp/service/SalesItemService.java
@@ -10,7 +10,7 @@ public interface SalesItemService {
 
     // 전체 상품 판매 내역 조회 (페이징/필터링 적용)
     // 반환 타입: Map<String, Object> (목록과 전체 개수를 포함)
-    Map<String, Object> getAllSalesItems(String startDate, String endDate, List<Long> itemIds, Long empNum, int page, int size); 
+    Map<String, Object> getAllSalesItems(String startDate, String endDate, List<Integer> itemIds, Integer empNum, int page, int size); 
 
     // 단일 상품 판매 내역 조회
     SalesItemDto getSalesItemById(Long itemSalesId);
@@ -35,8 +35,8 @@ public interface SalesItemService {
      */
     int deleteSalesItem(Long itemSalesId);
 
-	List<Map<String, Object>> getItemSalesAnalytics(String startDate, String endDate, List<Long> itemIds, Long memNum,
-			Long empNum);
+	List<Map<String, Object>> getItemSalesAnalytics(String startDate, String endDate, List<Integer> itemIds, Integer memNum,
+			Integer empNum);
 
 	Map<String, List<Map<String, Object>>> getItemSalesGraphData(String startDate, String endDate, String groupByUnit);
 }
